+++ library/src/main/java/cloud/pace/sdk/poikit/utils/PoiKitExceptions.kt
@@ -1,8 +1,8 @@
 package cloud.pace.sdk.poikit.utils
 
-class ApiException(var errorCode: Int, var errorMessage: String) : Exception() {
+class ApiException @JvmOverloads constructor(var errorCode: Int, var errorMessage: String, var requestId: String? = null) : Exception() {
     override fun toString(): String {
         return super.toString() + "\n" +
-            "code = $errorCode || message = $errorMessage"
+            "code = $errorCode || message = $errorMessage || request ID = $requestId"
     }
 }